Created more elegant work-around for CryptoStream closing/disposing our output stream.
Rather than copying to a temporary workspace/in-memory stream, we simply wrap our output stream in a special "PreventDispose" stream
such that when close/dispose are called on the stream (from within CryptoStream), the outer stream is not disposed.

The responsibility for closing the "output" stream is with the caller, not CryptoStream.